requirements.txt
requests
instagrapi
pillow